refactor: move to the styles folder
