+++ src/components/Header.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../App.css';
+import '../style/App.css';
 import { NavLink } from 'react-router-dom';
 import Location from './Location';
