+++ src/style/Form.css
@@ -3,15 +3,6 @@
   --red: #c03939;
 }
 
-.forms-page {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
 .forms-container {
   padding: 1rem;
   width: 60vw;
@@ -500,15 +491,4 @@ input[type=radio] {
   border: none;
   border-radius: 5px;
   background-color: var(--green);
-}
-
-.forms-page .container-cards {
-  width: 65vw;
-  padding: 2rem;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 2rem;
 }
